Generated ical files from yaml
So this would be one way to ensure we have proper calendars for the
different teams that people can import.

If we agree on where it lives and how it gets updated, I will write a
small README on how to use it (make changes, and import it into your
calendar).