Major revision of group-aws docs
With the move to collections we've made a number of changes.
In addition to new paths to issues, PRs, etc we're also implementing
a project board to track work that was historically manually tracked
in this repo (boto refactors, missing ci tests, etc).  Also adds
updates to best practices docs and other housekeeping updates.